requirements.macos.txt
amqp==2.5.1
Babel==2.4.0
backports-abc==0.5
billiard==3.6.1.0
celery==4.4.0rc3
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.18
#django-celery-beat==1.1.1
django-celery-beat==2.0.0
django-celery-results==1.1.2
django-cursor-pagination==0.1.4
django-datetime-widget==0.9.3
django-debug-toolbar==2.0
django-extensions==2.1.9
django-rest-swagger==2.2.0
django-timezone-field==4.0
djangorestframework==3.9.4
flower==0.9.2
future==0.16.0
futures==3.1.1
gunicorn==20.0.4
idna==2.5
importlib-metadata==0.23
itypes==1.1.0
Jinja2==2.11.3
jsonfield==2.0.2
kombu==4.6.4
libmagic==1.0
MarkupSafe==1.1.0
meld3==1.0.2
more-itertools==7.2.0
netaddr==0.7.19
openapi-codec==1.3.2
#psycopg2-binary==2.8.6
psycopg2-binary==2.7.5
python-crontab==2.3.6
python-dateutil==2.8.0
python-magic==0.4.15
python-magic-bin==0.4.14
python-memcached==1.59
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
singledispatch==3.4.0.3
six==1.10.0
sqlparse==0.3.0
supervisor==4.2.1
thehive4py==1.4.3
tornado==5.1
tzlocal==1.5.1
uritemplate==3.0.0
urllib3==1.25.9
vine==1.3.0
Werkzeug==0.15.3
zipp==0.6.0
django-filter==2.4.0
django-cors-headers==3.2.0
djangorestframework-simplejwt==4.4.0
django-health-check==3.12.1
psutil==5.7.0
django-organizations==1.1.2
django-annoying==0.10.6
django-reset-migrations
django-dbconn-retry
djangorestframework-datatables==0.5.2
